加入下雨下雪粒子 武道大会奖励配置文件修改

+++ Classes/GlobalData.cpp
@@ -1345,7 +1345,7 @@ void GlobalData::loadChallengeRewardData()
 		for (unsigned int m = 0; m < v.Size(); m++)
 		{
 			std::string str = v[m].GetString();
-			int rnd = atoi(str.c_str());
+			float rnd = atof(str.c_str());
 			if (str.length() > 0 && rnd > 0)
 			{
 				data.vec_winrnd.push_back(rnd);
